requirements.txt
asgiref==3.8.1
boto3==1.38.37
botocore==1.38.37
Django==5.2.3
djangorestframework==3.16.0
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
pytest
pytest-django
celery==5.3.4
amqp==5.1.1
django-celery-beat
django-celery-results
django-filter==23.5
spacy==3.7.6
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0


